+++ charlotte_core/src/framebuffer/colors.rs
@@ -3,6 +3,7 @@
 #[non_exhaustive]
 pub struct Color;
 
+#[allow(unused)]
 impl Color {
     pub const BLACK: u32 = 0x00000000;
     pub const WHITE: u32 = 0xFFFFFFFF;
@@ -17,7 +18,7 @@ impl Color {
 #[allow(unused)]
 pub fn blend_colors(foreground: u32, background: u32, blend_factor: u8) -> u32 {
     let fg_ratio = blend_factor as u32;
-    let bg_ratio = 255 - fg_ratio as u32;
+    let bg_ratio = 255 - fg_ratio;
 
     let r = (((foreground >> 16) & 0xFF) * fg_ratio + ((background >> 16) & 0xFF) * bg_ratio) / 255;
     let g = (((foreground >> 8) & 0xFF) * fg_ratio + ((background >> 8) & 0xFF) * bg_ratio) / 255;
